Use proper mode when opening temp files.
This commit fixes #2. SpooledTemporaryFiles start off as StringIO instances
and transition to on-disk temp files when the grow beyond 4M. As StringIO
instances, the opening mode is ignored. Once spooled to disk, the mode is
enforced.

Since this library chains output from functions to input of the next
function, reading from returned temporary files is required. Thus the mode
wb+ is appropriate.